backend/requirements.txt
# flask==2.3.2
# werkzeug==2.3.6
# flask-sqlalchemy==3.0.3
# flask-marshmallow==0.15.0
# marshmallow-sqlalchemy==0.29.0
# flask-cors==4.0.0
# flask-swagger-ui==4.11.1
# python-dotenv==1.0.0
# marshmallow==3.19.0
# SQLAlchemy==2.0.15

flask==2.3.2
flask-sqlalchemy==3.1.1
flask-marshmallow==0.15.0
marshmallow==3.20.1
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-swagger-ui==4.11.1